pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resumeagents"
version = "0.1.0"
description = "Multi-Agents LLM Resume & Job Application Framework"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "chai hyeon jong", email = "chaihyeonjong@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.40",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
vectordb = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.21.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/hjongc/ResumeAgents"
Repository = "https://github.com/hjongc/ResumeAgents"
Issues = "https://github.com/hjongc/ResumeAgents/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["resumeagents*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88 